BJS2-41432 - Анализ просмотров постов. Решил мердж конфликты, добавил логов.
