+++ src/main/java/faang/school/analytics/config/redis/RedisConfig.java
@@ -1,17 +1,21 @@
 package faang.school.analytics.config.redis;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import faang.school.analytics.listener.RecommendationEventListener;
 import faang.school.analytics.listener.postview.PostViewEventListener;
-import faang.school.analytics.service.analytic.AnalyticsEventService;
+import faang.school.analytics.service.events.AnalyticsEventService;
 import lombok.RequiredArgsConstructor;
+
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.RedisStandaloneConfiguration;
 import org.springframework.data.redis.connection.jedis.JedisConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.listener.ChannelTopic;
 import org.springframework.data.redis.listener.RedisMessageListenerContainer;
+import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
+import org.springframework.data.redis.serializer.StringRedisSerializer;
 
 @Configuration
 @RequiredArgsConstructor
@@ -20,21 +24,33 @@ public class RedisConfig {
     private final AnalyticsEventService analyticsEventService;
 
     @Value("${spring.data.redis.host}")
-    private String host;
+    private String redisHost;
     @Value("${spring.data.redis.port}")
-    private int port;
+    private int redisPort;
+
+    @Value("${${spring.data.redis.channel.recommendation_topic}")
+    private String recommendationChannel;
     @Value("${spring.data.redis.channel.post-view}")
     private String postViewTopicName;
 
     @Bean
-    public RedisConnectionFactory redisConnectionFactory() {
-        RedisStandaloneConfiguration connectionConfig = new RedisStandaloneConfiguration(host, port);
-        return new JedisConnectionFactory(connectionConfig);
+    public JedisConnectionFactory jedisConnectionFactory() {
+        RedisStandaloneConfiguration redisConfig = new RedisStandaloneConfiguration(redisHost, redisPort);
+        return new JedisConnectionFactory(redisConfig);
     }
 
     @Bean
-    public ChannelTopic postViewTopic() {
-        return new ChannelTopic(postViewTopicName);
+    public RedisTemplate<String, Object> redisTemplate() {
+        RedisTemplate<String, Object> template = new RedisTemplate<>();
+        template.setConnectionFactory(jedisConnectionFactory());
+        template.setKeySerializer(new StringRedisSerializer());
+        template.setValueSerializer(new StringRedisSerializer());
+        return template;
+    }
+
+    @Bean
+    public MessageListenerAdapter recommendationListener(RecommendationEventListener recommendationEventListener) {
+        return new MessageListenerAdapter(recommendationEventListener);
     }
 
     @Bean
@@ -43,10 +59,22 @@ public class RedisConfig {
     }
 
     @Bean
-    public RedisMessageListenerContainer redisContainer() {
+    public ChannelTopic recommendationTopic() {
+        return new ChannelTopic(recommendationChannel);
+    }
+
+    @Bean
+    public ChannelTopic postViewTopic() {
+        return new ChannelTopic(postViewTopicName);
+    }
+
+    @Bean
+    public RedisMessageListenerContainer redisContainer(MessageListenerAdapter recommendationListener) {
         RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(redisConnectionFactory());
+        container.setConnectionFactory(jedisConnectionFactory());
+        container.addMessageListener(recommendationListener, recommendationTopic());
         container.addMessageListener(postViewEventListener(), postViewTopic());
         return container;
     }
+
 }
