+++ src/main/java/faang/school/analytics/domain/dto/events/AnalyticsEventDto.java
@@ -1,6 +1,7 @@
 package faang.school.analytics.domain.dto.events;
 
 import jakarta.validation.constraints.NotNull;
+import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
