--- src/main/java/faang/school/analytics/dto/analytic/AnalyticsEventDto.java
@@ -1,28 +1,0 @@
-package faang.school.analytics.dto.analytic;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Data
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-public class AnalyticsEventDto {
-
-    @NotNull
-    private Long receiverId;
-
-    @NotNull
-    private Long actorId;
-
-    @JsonProperty(defaultValue = "-1")
-    private int eventTypeNumber;
-
-    private LocalDateTime receivedAt;
-}
