+++ src/main/java/faang/school/analytics/listener/RecommendationEventListener.java
@@ -1,11 +1,10 @@
 package faang.school.analytics.listener;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.analytics.dto.analytic.AnalyticsEventDto;
+import faang.school.analytics.domain.dto.events.AnalyticsEventDto;
 import faang.school.analytics.dto.recommendation.RecommendationEvent;
-import faang.school.analytics.mapper.AnalyticsEventMapper;
-import faang.school.analytics.model.EventType;
-import faang.school.analytics.service.analytic.AnalyticsEventService;
+import faang.school.analytics.mapper.events.AnalyticsEventMapper;
+import faang.school.analytics.service.events.AnalyticsEventService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.connection.Message;
