+++ src/main/java/faang/school/analytics/listener/postview/PostViewEventListener.java
@@ -2,12 +2,14 @@ package faang.school.analytics.listener.postview;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import faang.school.analytics.dto.analytic.AnalyticsEventDto;
-import faang.school.analytics.service.analytic.AnalyticsEventService;
+import faang.school.analytics.domain.dto.events.AnalyticsEventDto;
+import faang.school.analytics.service.events.AnalyticsEventService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.redis.connection.Message;
 import org.springframework.data.redis.connection.MessageListener;
 
+@Slf4j
 @RequiredArgsConstructor
 public class PostViewEventListener implements MessageListener {
     private final ObjectMapper objectMapper;
@@ -18,6 +20,7 @@ public class PostViewEventListener implements MessageListener {
         try {
             String json = new String(message.getBody());
             AnalyticsEventDto event = objectMapper.readValue(json, AnalyticsEventDto.class);
+            log.info("received the message: {}, send it to the service", event);
             analyticsEventService.savePostView(event);
         } catch (JsonProcessingException e) {
             throw new RuntimeException(e);
