+++ src/main/java/faang/school/analytics/mapper/events/AnalyticsEventMapper.java
@@ -1,23 +1,36 @@
 package faang.school.analytics.mapper.events;
 
 import faang.school.analytics.domain.dto.events.AnalyticsEventDto;
+import faang.school.analytics.dto.recommendation.RecommendationEvent;
 import faang.school.analytics.model.AnalyticsEvent;
 import faang.school.analytics.model.EventType;
 import org.mapstruct.Mapper;
+import org.mapstruct.Mapping;
+import org.mapstruct.Named;
 import org.mapstruct.ReportingPolicy;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface AnalyticsEventMapper {
-    AnalyticsEventDto toDto(AnalyticsEvent event);
+    EventType recommendation_received = EventType.RECOMMENDATION_RECEIVED;
 
-    AnalyticsEvent toEntity(AnalyticsEventDto eventDto);
+    @Mapping(source = "eventType", target = "eventType", qualifiedByName = "mapToEventType")
+    AnalyticsEvent toEntity(AnalyticsEventDto analyticsEventDto);
 
-    default Integer map(EventType type) {
-        return type.ordinal();
+    @Mapping(source = "eventType", target = "eventType", qualifiedByName = "mapToEventTypeNumber")
+    AnalyticsEventDto toDto(AnalyticsEvent analyticsEvent);
+
+    @Mapping(source = "authorId", target = "actorId")
+    @Mapping(source = "createdAt", target = "receivedAt")
+    @Mapping(target = "eventType", expression = "java(map(recommendation_received))")
+    AnalyticsEventDto recommendationToAnalyticsDto(RecommendationEvent recommendationEvent);
+
+    @Named("mapToEventTypeNumber")
+    default int map(EventType eventType) {
+        return eventType.ordinal();
     }
 
-    default EventType map(Integer value) {
-        return EventType.of(value);
+    @Named("mapToEventType")
+    default EventType map(int eventType) {
+        return EventType.of(eventType);
     }
-
 }
