+++ src/main/java/faang/school/analytics/scheduler/analytic/ScheduledRatingUpdater.java
@@ -5,7 +5,7 @@ import faang.school.analytics.config.context.UserContext;
 import faang.school.analytics.mapper.user.UpdateUsersRankMapper;
 import faang.school.analytics.model.EventType;
 import faang.school.analytics.repository.analytic.AnalyticsEventRepository;
-import faang.school.analytics.service.analytic.AnalyticsEventService;
+import faang.school.analytics.service.events.AnalyticsEventService;
 import faang.school.analytics.service.analytic.AverageValueOfActionCalculator;
 import faang.school.analytics.service.analytic.StandardDeviationCalculator;
 import lombok.RequiredArgsConstructor;
