+++ src/main/java/faang/school/analytics/service/events/AnalyticsEventService.java
@@ -1,5 +1,7 @@
 package faang.school.analytics.service.events;
 
+import faang.school.analytics.client.user.UserServiceClient;
+import faang.school.analytics.config.context.UserContext;
 import faang.school.analytics.domain.dto.events.AnalyticsEventDto;
 import faang.school.analytics.domain.dto.events.AnalyticsEventFilterDto;
 import faang.school.analytics.exception.DataValidationException;
@@ -8,11 +10,15 @@ import faang.school.analytics.model.AnalyticsEvent;
 import faang.school.analytics.repository.analytic.AnalyticsEventRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 @Slf4j
@@ -22,6 +28,8 @@ public class AnalyticsEventService {
     private final AnalyticsEventFilter analyticsEventFilter;
     private final AnalyticsEventMapper analyticsEventMapper;
     private final AnalyticsEventRepository analyticsEventRepository;
+    private final UserServiceClient userServiceClient;
+    private final UserContext userContext;
 
     public AnalyticsEventDto saveEvent(AnalyticsEventDto eventDto) {
         if (eventDto.getId() != null) {
@@ -49,11 +57,41 @@ public class AnalyticsEventService {
     }
 
     private Stream<AnalyticsEvent> filterEvents(Stream<AnalyticsEvent> events, AnalyticsEventFilterDto filter) {
-       if (filter.getInterval() == null && filter.getFrom() == null && filter.getTo() == null) {
-           return events;
-       }
-       return filter.getInterval() != null ? analyticsEventFilter.filterByInterval(events, filter.getInterval().getDays()) :
-               analyticsEventFilter.filterByDates(events, filter.getFrom(), filter.getTo());
+        if (filter.getInterval() == null && filter.getFrom() == null && filter.getTo() == null) {
+            return events;
+        }
+        return filter.getInterval() != null ? analyticsEventFilter.filterByInterval(events, filter.getInterval().getDays()) :
+                analyticsEventFilter.filterByDates(events, filter.getFrom(), filter.getTo());
     }
 
+    public ResponseEntity<Void> saveAction(AnalyticsEventDto analyticsEventDto) {
+        userContext.setUserId(1L);
+        userServiceClient.getUser(analyticsEventDto.getReceiverId());
+        analyticsEventDto.setReceivedAt(LocalDateTime.now());
+        log.info("getting action: {}, from userId: {}", analyticsEventDto, analyticsEventDto.getReceiverId());
+        analyticsEventRepository.save(analyticsEventMapper.toEntity(analyticsEventDto));
+        log.info("success saved action of userId: {}", analyticsEventDto.getReceiverId());
+        return ResponseEntity.ok().build();
+    }
+
+    public Map<Long, Integer> mapAnalyticEventsToActorActionsCount(List<AnalyticsEvent> analyticsEvents) {
+        return analyticsEvents.stream()
+                .collect(Collectors.groupingBy(
+                        AnalyticsEvent::getActorId,
+                        HashMap::new,
+                        Collectors.summingInt(event -> 1)
+                ));
+    }
+
+    public int getSumOfUsersActionsByEventType(List<Integer> usersActionsSumByEventType) {
+        return usersActionsSumByEventType.stream()
+                .mapToInt(Integer::intValue)
+                .sum();
+    }
+
+    public AnalyticsEventDto savePostView(AnalyticsEventDto analyticsEventDto) {
+        log.info("saving view, post id: {}", analyticsEventDto.getReceiverId());
+        AnalyticsEvent event = analyticsEventMapper.toEntity(analyticsEventDto);
+        return analyticsEventMapper.toDto(analyticsEventRepository.save(event));
+    }
 }
